refactor: reduce code duplication in pkg/sources
